YandexGPTBot/requirements.txt
requests==2.32.5
fastapi==0.116.2
uvicorn==0.35.0
pyjwt==2.10.1
python-dotenv==1.1.1
langchain-community==0.3.29
langchain==0.3.27
yandex_cloud_ml_sdk==0.15.0
numpy==2.0.0
scipy
